Fix electra pretraining
